Small bugfix and more bugs
Bugfixed and error where result of mul wasn't printed correctly if Num2 was a little bit longer than it is supposed.

Multiplication is still not working as it was intended -> you can multiply by more than one number.

And it seems you can't multiply using letters from B to F as it messes up with conversion .... somehow.